+++ src/components/content/care-team/patient-careteam.tsx
@@ -1,7 +1,7 @@
 import cx from "classnames";
 import { useRef } from "react";
-import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
 import { patientCareTeamColumns } from "./patient-careteam-columns";
+import { useResourceDetailsDrawer } from "../resource/resource-details-drawer";
 import { useCTW } from "@/components/core/providers/ctw-provider";
 import { Table } from "@/components/core/table/table";
 import { usePatientCareTeam } from "@/fhir/care-team";
@@ -29,6 +29,7 @@ export function PatientCareTeam({ className }: PatientCareTeamProps) {
     header: (m) => m.practitionerName,
     subHeader: (m) => m.qualification,
     details: careTeamData,
+    getSourceDocument: true,
   });
 
   return (