+++ src/components/content/conditions/helpers/details.tsx
@@ -1,13 +1,12 @@
-import { useResourceDetailsDrawer } from "../../resource/resource-details-drawer";
 import { useConditionHistory } from "./history";
 import { useConfirmDeleteCondition, useEditConditionForm } from "./modal-hooks";
+import { useResourceDetailsDrawer } from "../../resource/resource-details-drawer";
 import { NotesList } from "@/components/core/notes-list";
 import { ConditionModel } from "@/fhir/models";
 import { capitalize } from "@/utils/nodash";
 
 export const useConditionDetailsDrawer = ({
   canEdit,
-  canRemove,
 }: {
   canEdit: boolean;
   canRemove: boolean;