+++ src/components/content/medications/story-helpers/mocks/requests.ts
@@ -46,7 +46,11 @@ function mockRequests() {
       if (req.url.searchParams.get("firstparty")) {
         return res(ctx.status(200), ctx.json(patientProviderMedsCache));
       }
-      return res(ctx.status(200), ctx.json(patientOtherProviderMedsCache));
+      return res(
+        ctx.delay(750),
+        ctx.status(200),
+        ctx.json(patientOtherProviderMedsCache)
+      );
     }
   );